+++ app/src/main/java/io/github/jbellis/brokk/difftool/ui/BrokkDiffPanel.java
@@ -10,6 +10,7 @@ import io.github.jbellis.brokk.difftool.node.JMDiffNode;
 import io.github.jbellis.brokk.difftool.performance.PerformanceConstants;
 import io.github.jbellis.brokk.git.GitRepo;
 import io.github.jbellis.brokk.gui.Chrome;
+import io.github.jbellis.brokk.gui.GitUiUtil;
 import io.github.jbellis.brokk.gui.GuiTheme;
 import io.github.jbellis.brokk.gui.ThemeAware;
 import io.github.jbellis.brokk.gui.util.KeyboardShortcutUtil;
@@ -20,7 +21,6 @@ import java.awt.event.KeyEvent;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.regex.Pattern;
 import javax.swing.*;
 import javax.swing.event.AncestorEvent;
 import javax.swing.event.AncestorListener;
@@ -506,8 +506,8 @@ public class BrokkDiffPanel extends JPanel implements ThemeAware {
             }
             var description = "Captured Diff: %s vs %s"
                     .formatted(
-                            friendlyCommitLabel(currentLeftSource.title(), repo),
-                            friendlyCommitLabel(currentRightSource.title(), repo));
+                            GitUiUtil.friendlyCommitLabel(currentLeftSource.title(), repo),
+                            GitUiUtil.friendlyCommitLabel(currentRightSource.title(), repo));
 
             var patch = DiffUtils.diff(leftLines, rightLines, (DiffAlgorithmListener) null);
             var unifiedDiff = UnifiedDiffUtils.generateUnifiedDiff(
@@ -1023,62 +1023,6 @@ public class BrokkDiffPanel extends JPanel implements ThemeAware {
         updateNavigationButtons();
         revalidate();
         repaint();
-    }
-
-    /**
-     * Builds a concise commit label such as<br>
-     * 'First line …' [abcdef1] <br>
-     * If {@code rawTitle} is just a hash, we try to look up the commit message from {@code repo}. When the lookup fails
-     * we fall back to the hash only.
-     */
-    private static String friendlyCommitLabel(@Nullable String rawTitle, @Nullable GitRepo repo) {
-        if (rawTitle == null || rawTitle.isBlank()) {
-            return "";
-        }
-
-        // Detect a leading full hash (40 hex chars) or short hash (≥7 hex chars)
-        var matcher =
-                Pattern.compile("^(?<hash>[a-fA-F0-9]{7,40})\\s*(?<msg>.*)$").matcher(rawTitle.strip());
-        String hash;
-        String msg;
-        if (matcher.matches()) {
-            hash = matcher.group("hash");
-            msg = matcher.group("msg");
-        } else {
-            // Try to find the hash at the end in brackets, e.g. "Some title (abcdef123)"
-            var tailMatcher = Pattern.compile("^(?<msg>.*)\\s+\\((?<hash>[a-fA-F0-9]{7,40})\\)$")
-                    .matcher(rawTitle.strip());
-            if (tailMatcher.matches()) {
-                hash = tailMatcher.group("hash");
-                msg = tailMatcher.group("msg");
-            } else {
-                // No recognisable hash – just truncate the string we have
-                return truncateWithEllipsis(rawTitle.trim(), 40);
-            }
-        }
-
-        hash = hash.substring(0, 7); // canonical short hash
-
-        // If we still have no message, and a repo is available, try to look it up
-        if ((msg == null || msg.isBlank()) && repo != null) {
-            try {
-                var infoOpt = repo.getLocalCommitInfo(hash);
-                if (infoOpt.isPresent()) {
-                    msg = infoOpt.get().message();
-                }
-            } catch (Exception ignore) {
-                /* lookup failure is non-fatal */
-            }
-        }
-
-        // Final formatting
-        msg = (msg == null ? "" : truncateWithEllipsis(msg.split("\\R", 2)[0].trim(), 30));
-        return msg.isBlank() ? "[%s]".formatted(hash) : "'%s' [%s]".formatted(msg, hash);
-    }
-
-    /** Truncates s to maxLen characters, appending '...' if needed. */
-    private static String truncateWithEllipsis(String s, int maxLen) {
-        return s.length() <= maxLen ? s : s.substring(0, maxLen - 2) + "...";
     }
 
     private void refreshAllDiffPanels() {
